homestream/requirements.txt
Django==2.2.9
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
python-magic==0.4.15
pytz==2019.3
sqlparse==0.3.0
